Disable the Trans2 pad2 field sometimes
Disable the Trans2 pad2 field when there is no data to pad. This fixes
an issue when parsing certain requests where the pad would require data
and corrupt the field before it.

Also update unit tests to account for the fact that the pad is only
present if there is data to pad.