Adjust how and when dynamic files are generated
This switches virtual dynamic files to be generated and converted to
static files when the share is opened. This means that the size does not
need to be known and that the contents can be scoped to the client and
session. The down side is that the contents do not change each time the
file is read but rather each time the share has been connected.